% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{calc_gxg}
\alias{calc_gxg}
\title{Calculate GxG's of HydroMonitor ObservationWell data.}
\usage{
calc_gxg(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}
}
\value{
Characteristics of monitoring well (meta data, \code{\link{read_export_csv}}) with the following fields added:
\itemize{
\item AHG & AHG_MV 99,85 \% value of observed heads (relative to REF, relative to soil surface level).
\item MHG & MHG_MV 97,7 \% value of observed heads (relative to REF, relative to soil surface level).
\item GHG & GHG_MV  84,1 \% value of observed head (relative to REF, relative to soil surface level).
\item GG & GG_MV 50\% value of observed heads (relative to REF, relative to soil surface level).
\item GLG & GLG_MV 15,9\% value of observed heads (relative to REF, relative to soil surface level).
\item MLG & MLG_MV 2,3\% value of observed heads (relative to REF, relative to soil surface level).
\item ALG & ALG_MV 0,15\% value of observed heads (relative to REF, relative to soil surface level).
\item n Number of observations used to calculate percentile values.
}
}
\description{
Percentile values of observed groundwater heads are calculated according to:
}
\details{
GxG's of observations of every filter is calculated.
}
\examples{
\dontrun{
gxg <- hm1 |> calc_gxg()
}
}
\seealso{
\code{link{gxg_table}}
\href{https://edepot.wur.nl/175881}{'Een alternatieve GHG analyse' Drs. D.H. Edelman, Ir. A.S. Burger
Stromingen 15 (2009) nummer 3 p29-34.}
}
