% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{create_shp}
\alias{create_shp}
\title{Create a point shape file from HydroMonitor ObservationWell data object.}
\usage{
create_shp(hm, crs = "EPSG:28992")
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}

\item{crs}{(character). Defaults to "EPSG:28992"}
}
\value{
SpatVector
}
\description{
Create a summarizing point shape file from the meta data part of HydroMonitor ObservationWell data object.
}
\details{
(\code{\link{read_export_csv}}).

Fields in the attribute table of the resulting shapefile are:
\itemize{
\item NAME: Name of observationwell.
\item X: X-coordinate of the observationwell.
\item Y: Y-coordinate of the observationwell.
\item NFILTERS: Number of filters.
\item TOP: Level of the top of the highest filter.
\item BOT: Level of the bottom of the lowest filter.
\item MV: Surface level.
}
}
\examples{
\dontrun{
v <- hm1 |> create_shp()
}
}
