% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{filter_on_extent}
\alias{filter_on_extent}
\title{Filter HydroMonitor ObservationWell data on extent.}
\usage{
filter_on_extent(hm, e)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}

\item{e}{Extent object or a Raster* or Spatial* object}
}
\value{
Filtered HydroMonitor ObservationWell data (list of 2 met data.frames xm en xd).
}
\description{
Filter HydroMonitor ObservationWell data on extent.
}
\examples{
\dontrun{
hm <- hm1
xmin <- min(hm$xm$X)
xmax <- max(hm$xm$X)
ymin <- min(hm$xm$Y)
ymax <- max(hm$xm$Y)
dx <- (xmax-xmin)/10
dy <- (ymax-ymin)/10
e <- terra::ext(c(xmin+dx, xmax-dx, ymin+dy, ymax-dy))
hm_filtered <- hm |> filter_on_extent( e )
}
}
