% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{filter_on_poly}
\alias{filter_on_poly}
\title{Filter HydroMonitor ObservationWell data with polygon shape.}
\usage{
filter_on_poly(hm, p, crs = "EPSG:28992")
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}

\item{p}{Polygon shape}

\item{crs}{(character). Defaults to "EPSG:28992"}
}
\value{
Filtered HydroMonitor ObservationWell data (list of 2 met data.frames xm en xd).
}
\description{
Filter HydroMonitor ObservationWell data with polygon shape.
}
\details{
Only polygon shapes of length=1 can be used.
}
\examples{
\dontrun{
hm <- hm1
p <- system.file("extdata","polygn.shp",package="hydromonitor") |> terra::vect()
hm_filtered_on_polygon <- filter_on_poly(hm, p)
}
}
