% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{filter_on_year}
\alias{filter_on_year}
\title{Filter HydroMonitor ObservationWell data on year.}
\usage{
filter_on_year(hm, minyear = 1900, maxyear = 3000)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}

\item{minyear}{Minimal year to read data from (integer)}

\item{maxyear}{Maximal year to read data from (integer)}
}
\value{
Filtered HydroMonitor ObservationWell data
}
\description{
Filter HydroMonitor ObservationWell data on year.
}
\examples{
\dontrun{
filtered <- hm1 |> filter_on_year( minyear=2000)
}
}
