% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{gxg_table}
\alias{gxg_table}
\title{Calculate GxG table based on head observations in top filters of HydroMonitor ObservationWell data.}
\usage{
gxg_table(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}
}
\value{
Tibble with parameters as specified in \code{link{calc_gxg}}.
}
\description{
Calculate GxG table based on head observations in top filters of HydroMonitor ObservationWell data.
}
\examples{
\dontrun{
hm <- hm1
gxg <- gxg_table(hm)
}
}
\seealso{
\code{link{calc_gxg}}
}
