% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm1}
\alias{hm1}
\title{HydroMonitor Observation Well Data}
\format{
An object of class \code{list} of length 2.
}
\usage{
hm1
}
\description{
Example dataset 1 to be used in examples,
}
\keyword{datasets}
