% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hm2}
\alias{hm2}
\title{HydroMonitor Observation Well Data}
\format{
An object of class \code{list} of length 2.
}
\usage{
hm2
}
\description{
Example dataset 2 to be used in examples
}
\keyword{datasets}
