% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{hm_plot}
\alias{hm_plot}
\title{Plot HydroMonitor ObservationWell data and optionally save plots to specified folder.}
\usage{
hm_plot(hm, output_dir = NULL)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}

\item{output_dir}{folder name to write plot (character). If NULL, plots are not written do disk.}
}
\value{
tibble. Fields:
\itemize{
\item NAME Name of observationwell (character vector)
\item plots List of timeseries plots (ggplot object)
}
}
\description{
Create a list of timeseries plots of all HydroMonitor Observationwell data
(\code{\link{read_export_csv}}).
}
\examples{
\dontrun{
x <- hm1 |> hm_plot()
NAME <- x$NAME[3]
x$plots[[which(x$NAME==NAME)]]
}
}
