% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydromonitor.R
\name{hydromonitor}
\alias{hydromonitor}
\title{hydromonitor: A package read / write / manipulate \href{https://www.kwrwater.nl/tools-producten/hydromonitor/}{HydroMonitor} Observationwell data.}
\description{
This package exports the following sample data sets/objects:
\itemize{
\item hm1, hm2: HydroMonitor ObservationWell Data. Can be read with \code{\link{read_export_csv}}.
\item polygn: Polygon shape used in examples.
}
}
\details{
This package exports the following functions:
\itemize{
\item \code{\link{read_export_csv}}
\item \code{\link{read_export_csv2}}
\item \code{\link{read_dino}}
\item \code{\link{read_dino_path}}
\item \code{\link{read_dino_zip}}
\item \code{\link{obs_periods}}
\item \code{\link{filter_on_year}}
\item \code{\link{filter_on_extent}}
\item \code{\link{filter_on_poly}}
\item \code{\link{rm_fltrs_with_no_obs}}
\item \code{\link{hm_rbind}}
\item \code{\link{rm_dble_fltrs}}
\item \code{\link{rm_dble_obs}}
\item \code{\link{hm_plot}}
\item \code{\link{calc_gxg}}
\item \code{\link{create_shp}}
\item \code{\link{nr_obs_ratio}}
}
}
