% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{nr_obs_ratio}
\alias{nr_obs_ratio}
\title{Ratio's (# observations in filter) / (average # of observations in monitoring well)}
\usage{
nr_obs_ratio(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}
}
\value{
tible with fields:
\itemize{
\item NAME Name of observationwell (character vector)
\item FILTER Filter number (integer)
\item RATIO  Ratio (# observations in filter) / (average # of observations in monitoring well) (numeric)
}
}
\description{
Ratio's (# observations in filter) / (average # of observations in monitoring well)
}
\examples{
\dontrun{
r <- hm1 |> nr_obs_ratio()
}
}
