% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{obs_periods}
\alias{obs_periods}
\title{Create a data frame with observation period for each filter.}
\usage{
obs_periods(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}
}
\value{
Data frame with fields "min_date" and "max_date" (POSIXct)
}
\description{
Create a data frame with observation period for each filter.
}
\examples{
\dontrun{
obs_periods( hm1 )
}
}
