% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bro_zip.R
\name{read_bro_zip}
\alias{read_bro_zip}
\title{Read all BROloket export files with measured heads from zip file.}
\usage{
read_bro_zip(fname, id = "nitgCode")
}
\arguments{
\item{fname}{Filename (character)}

\item{id}{Code to use in the column 'NAME' of the resulting data.frame (character).}
}
\value{
List of 2:
\itemize{
\item xm Characteristics of the monitoring well (meta data).
\item xd Measured heads (data.frame).
}

Variables in data.frame xm:
\itemize{
\item NAME Name of observationwell.  (character)
\item FILTER Filter number (integer)
\item X x-coordinate of observationwell (numeric)
\item Y y-coordinate of observationwell (numeric)
\item TOP Level of the top of filter (m+REF, numeric)
\item BOT Level of the bottom of filter (m+REF,numeric)
\item MV Surface level (m+REF,numeric)
}

Variables in data.frame xd:
\itemize{
\item NAME Name of observationwell (character)
\item FILTER Filter number (integer)
\item DATE Date of observation (POSIXct)
\item HEAD Observed head (numeric)
}
}
\description{
Read all BROloket export files with measured heads from zip file.
}
\details{
id is one of "broId", "nitgCode", "wellCode", "brold".
}
\examples{
hm1 <- system.file("extdata","bro_export.zip",package="hydromonitor") |> read_bro_zip( )
}
