% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dino.R
\name{read_dino}
\alias{read_dino}
\title{Read export Dino *_1.csv file with measured heads.}
\usage{
read_dino(fname)
}
\arguments{
\item{fname}{Filename (csv file, character)}
}
\value{
list met twee items: xm (data.frame) met kopgegevens and xd (data.frame) met meetgegevens (data.frame).
}
\description{
Trailing letters of observation well are removed.
}
\examples{
hm3 <- system.file("extdata","B38E0108001_1.csv",package="hydromonitor") |> read_dino( )
}
