% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dino.R
\name{read_dino_path}
\alias{read_dino_path}
\title{Read all export Dino *_1.csv files with measured heads in specified folder.}
\usage{
read_dino_path(path)
}
\arguments{
\item{path}{A character vector containing a single path name. Tilde expansion (see \code{\link{path.expand}}) is done.}
}
\description{
Read all export Dino *_1.csv files with measured heads in specified folder.
}
\examples{
hm4 <- system.file("extdata","Grondwaterstanden_Put",package="hydromonitor") |> read_dino_path( )
}
