% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_export_csv.R
\name{read_export_csv}
\alias{read_export_csv}
\title{Read export HydroMonitor file with ObservationWell data.}
\usage{
read_export_csv(fname)
}
\arguments{
\item{fname}{Filename (csv file, character)}
}
\value{
List of 2:
\itemize{
\item xm Characteristics of the monitoring well (meta data).
\item xd Measured heads (data.frame).
}

Variables in data.frame xm:
\itemize{
\item NAME Name of observationwell.  (character)
\item FILTER Filter number (integer)
\item X x-coordinate of observationwell (numeric)
\item Y y-coordinate of observationwell (numeric)
\item TOP Level of the top of filter (m+REF, numeric)
\item BOT Level of the bottom of filter (m+REF,numeric)
\item MV Surface level (m+REF,numeric)
}

Variables in data.frame xd:
\itemize{
\item NAME Name of observationwell (character)
\item FILTER Filter number (integer)
\item DATE Date of observation (POSIXct)
\item HEAD Observed head (numeric)
}
}
\description{
Trailing letters of observation well are removed.
}
\examples{
hm1 <- system.file("extdata","export_hm.csv",package="hydromonitor") |> read_export_csv( )
}
