% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_export_csv.R
\name{remove_trailing_letter}
\alias{remove_trailing_letter}
\title{Remove trailing letter of a string.}
\usage{
remove_trailing_letter(x)
}
\arguments{
\item{x}{A character vector where matches are sought, or an object which can be coerced
by as.character to a character vector. Long vectors are supported.}
}
\value{
A character vector of the same length and with the same attributes as x.
}
\description{
Remove trailing letter of a string.
}
\examples{
\dontrun{remove_trailing_letter("13250900B002A")}
}
