% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{rm_dble_fltrs}
\alias{rm_dble_fltrs}
\title{Remove double filter information in meta data part of HydroMonitor ObservationWell data}
\usage{
rm_dble_fltrs(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}
}
\value{
HydroMonitor ObservationWell data with double filter information removed from meta data part of
HydroMonitor ObservationWell data \code{\link{read_export_csv}}
}
\description{
Remove double filter information in meta data part of HydroMonitor ObservationWell data
}
\examples{
\dontrun{
hm_filtered <- hm1 |> rm_dble_fltrs( )
}
}
