% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{rm_dble_obs}
\alias{rm_dble_obs}
\title{Remove double observations in HydroMonitor ObservationWell data.}
\usage{
rm_dble_obs(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}
}
\value{
HydroMonitor ObservationWell data with double records removed.
}
\description{
Remove double observations in HydroMonitor ObservationWell data.
}
\examples{
\dontrun{
hm_filtered <- hm1 |> rm_dble_obs( )
}
}
