% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitys.R
\name{rm_fltrs_with_no_obs}
\alias{rm_fltrs_with_no_obs}
\title{Remove meta data of filters if there are no observations available.}
\usage{
rm_fltrs_with_no_obs(hm)
}
\arguments{
\item{hm}{HydroMonitor Observation Well data as read by \code{\link{read_export_csv}}}
}
\value{
Filtered HydroMonitor ObservationWell data (list of 2 met data.frames xm en xd).
}
\description{
Remove meta data of filters if there are no observations available.
}
\examples{
\dontrun{
hm_filtered <- hm1 |> rm_fltrs_with_no_obs( )
}
}
